Move routes into lib/constants
